POC - Use configurable Timer to wake up at configurable duration
MVP Demo of using Timer1 to wake up from Idle at a user-configurable
time with high accuracy for times less than 8.3 seconds.

So far, it only lets us use idle, so we can get 17uA - which, to be
fair, is better than 30.

I should put this solution on the github!