Experimental Mode: Dynamically moving V2 for precise timing
On the way down, we threshold at 1.65 on whether we should wake up or
not.

BUt, on the way up, we should threshold exactly at 2.0 which is exactly
the zero-crossing point.  Else, we are turning the coil on while hte
magnet is coming in (For a few hundred uS), thus, robbing power.

This works... but is it a good idea or not?  That will take some
testing.  Which I will do TOMORROW.  Testing results will show up in a
second commit and comments therof.