Develop a new (complicated) mode to measure interrupt_in timing
Requires leaving Timer1 running and idling, not powering_down, but it's
possilble to measure the pulse duration.

We can power_down for when the pulse is low (not demonstrated here), and
then only do Idle when we the pulse is HIGH, so we accurately measure
the HIGH time but not the low time.

We _do_ use a higher power setting mode (Lighsleep) - but unfortunately,
it is power standby, not idle.  Power standby takes approx 500uA, idle
meanwhile is like 15mA. Idle is still better than no idle, AND, we would
be only idling for like ... 15 ms?

But if I can figure out a way to pull this off with timer 2 - not timer
1 - we can leave Timer 2 on during power standby.